+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "bevy_vrm1"
-version = "0.1.2"
+version = "0.2.0-dev"
 edition = "2021"
 authors = ["notelm <elmprograminfo@gmail.com>"]
 description = "Allows you to use VRM and VRMA in Bevy"
@@ -23,7 +23,7 @@ serde = "1"
 serde_json = "1"
 anyhow = "1"
 bitflags = { version = "2.9" }
-nonmax = "0.5.5"
+nonmax = "0.5"
 
 [dev-dependencies]
 bevy = { version = "0.16.0" }
