+++ src/vrm/gltf.rs
@@ -1,14 +1,9 @@
 pub mod extensions;
 pub mod materials;
 
-pub mod prelude{
+pub mod prelude {
     pub use crate::vrm::gltf::{
-        extensions::{
-            VrmExtensions,
-            VrmNode,
-            vrmc_vrm::*,
-            vrmc_spring_bone::*,
-        },
+        extensions::{vrmc_spring_bone::*, vrmc_vrm::*, VrmExtensions, VrmNode},
         materials::*,
     };
 }
