+++ src/vrm/mtoon/material/outline.rs
@@ -1,21 +1,24 @@
 use crate::prelude::*;
 use bevy::asset::Asset;
 use bevy::color::LinearRgba;
-use bevy::prelude::{Component, Reflect, TypePath};
-use bevy::render::extract_component::ExtractComponent;
+use bevy::prelude::{Reflect, TypePath};
 
 #[cfg_attr(feature = "reflect", derive(Reflect))]
 #[cfg_attr(not(feature = "reflect"), derive(TypePath))]
-#[derive(Asset, PartialEq, Debug, Clone, Component, ExtractComponent, Default)]
+#[derive(Asset, PartialEq, Debug, Clone, Default)]
 pub struct MToonOutline {
+    /// [`OutlineWidthMode`]
     pub mode: OutlineWidthMode,
     /// The outline width
     ///
-    /// the unit is in meters.
+    /// the unit is in meters.  
+    /// The outline width.
     pub width_factor: f32,
 
+    /// The outline color.
     pub color: LinearRgba,
 
+    /// The factor for the outline lighting mix.
     pub lighting_mix_factor: f32,
 }
 
@@ -34,6 +37,7 @@ impl From<&VrmcMaterialsExtensitions> for MToonOutline {
     }
 }
 
+/// Please see [`outlinewidthmode`](https://github.com/vrm-c/vrm-specification/blob/master/specification/VRMC_materials_mtoon-1.0/README.md#outlinewidthmode) for details.
 #[derive(Reflect, Debug, Clone, Default, Copy, PartialEq, Eq, Hash)]
 pub enum OutlineWidthMode {
     /// The outline will not be drawn.
